Consume new RGs api package
