+++ utils/src/pickTreeItem/quickPickAzureResource/QuickPickGroupStep.ts
@@ -3,15 +3,15 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as types from '../../../index';
 import * as vscode from 'vscode';
 import { GenericQuickPickStep, SkipIfOneQuickPickOptions } from '../GenericQuickPickStep';
-import { AzureResourceQuickPickWizardContext } from '../../../hostapi.v2';
+import { AzureResourceQuickPickWizardContext } from '../../../index';
 import { parseContextValue } from '../../utils/contextUtils';
 import { PickFilter } from '../PickFilter';
+import type { AzExtResourceType } from '@microsoft/vscode-azureresources-api';
 
 interface GroupQuickPickOptions extends SkipIfOneQuickPickOptions {
-    groupType?: types.AzExtResourceType[];
+    groupType?: AzExtResourceType[];
     skipIfOne?: true;
 }
